clean up power spectrum methods
